made component for create/edit modal
