+++ post-app/src/components/CreateEditPostModal/CreateEditPostModalStyle.module.css
@@ -1,0 +1,59 @@
+.createPostModal {
+    display: block;
+    margin: 40px 25px 0;
+    padding: 10px 15px;
+    position: initial;
+    background-color: rgb(73, 71, 71);
+    border-radius: 10px;
+    color: rgb(221, 225, 228);
+  }
+
+  .titleAndXButton {
+    display: grid;
+    grid-template-columns: 15fr 1fr;
+    column-gap: 20px;
+    justify-content: end;
+  }
+
+  .title {
+    justify-self: center;
+  }
+
+  .userAndDropDowns {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .userAndDropDown {
+    margin-left: 2px;
+    margin-bottom: 8px;
+  }
+
+  .username a {
+    display: block;
+    width: 67px;
+  }
+
+  .username {
+    min-width: 155px;
+    padding-top: 5px;
+    margin-bottom: 3px;
+    font-weight: bold;
+    font-size: 18px;
+  }
+
+  .dropdownButton {
+    margin-bottom: 20px;
+  }
+
+  .textArea {
+    margin-bottom: 20px;
+    margin-top: -2px;
+    background-color: rgb(217, 225, 233);
+    border: none;
+    font-size: 18px;
+  }
+
+  .submitBtn {
+    width: 100%;
+  }
