+++ post-app/src/components/CreatePost/CreatePost.js
@@ -4,19 +4,18 @@ import { NavLink, useNavigate } from "react-router-dom";
 
 import {
   optionValues as privacyValues,
-  privacyOptions,
 } from "./privacyOptions";
 import { optionValues as categoryValues } from "./categoryOptions";
-import { categoryOptions } from "./categoryOptions";
 import { addPost, getPosts } from "../../store/actions/post/postActions";
+import CreateEditPostModal from "../CreateEditPostModal/CreateEditPostModal";
 
-import { Button, Dropdown, DropdownButton, Form, Modal } from "react-bootstrap";
+import { Button } from "react-bootstrap";
 import styles from "./CreatePostStyle.module.css";
 
 function CreatePost() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const [showModal, setShowModal] = useState(false); 
+  const [showModal, setShowModal] = useState(false);
   const [privacy, setPrivacy] = useState({
     value: privacyValues.PRIVATE,
     label: "Private",
@@ -30,13 +29,6 @@ function CreatePost() {
     (state) => state.authReducer.isAuthenticated
   );
   const user = useSelector((state) => state.authReducer.userInfo);
-  const handlePrivacy = (option) => {
-    setPrivacy(option);
-  };
-
-  const handleCategory = (option) => {
-    setCategory(option);
-  };
 
   const onPost = () => {
     content.trim();
@@ -75,7 +67,6 @@ function CreatePost() {
 
   return (
     <>
-
       {!showModal && (
         <div className={styles.createPost}>
           <div className={isAuthenticated ? styles.username : styles.sighIn}>
@@ -96,87 +87,17 @@ function CreatePost() {
       )}
 
       {showModal && (
-        <div
-          className={`${styles.modal} ${styles.show} ${styles.createPostModal}`}
-        >
-          <Modal.Dialog>
-            <div className={styles.titleAndXButton}>
-              <Modal.Header className={styles.title}>
-                <Modal.Title>Create post</Modal.Title>
-              </Modal.Header>
-
-              <Button variant="secondary" onClick={() => setShowModal(false)}>
-                X
-              </Button>
-            </div>
-            <hr />
-            <div className={styles.userAndDropDowns}>
-              <div className={styles.userAndDropDown}>
-                <div className={styles.username}>
-                  {user && `${user.name} ${user.surname}`}
-                </div>
-
-                <DropdownButton
-                  className={styles.dropdownButton}
-                  variant="primary"
-                  title={privacy.value ? privacy.label : "Private"}
-                  id="privacy"
-                >
-                  {privacyOptions.map((option, index) => (
-                    <Dropdown.Item
-                      className={styles.dropdownItem}
-                      key={index}
-                      active={privacy.value === option.value}
-                      onClick={() => handlePrivacy(option)}
-                    >
-                      {option.label}
-                    </Dropdown.Item>
-                  ))}
-                </DropdownButton>
-              </div>
-              <div>
-                <DropdownButton
-                  className={styles.dropdownButton}
-                  variant="primary"
-                  title={category.value ? category.label : "General"}
-                  id="category"
-                >
-                  {categoryOptions.map((option, index) => (
-                    <Dropdown.Item
-                      className={styles.dropdownItem}
-                      key={index}
-                      active={category.value === option.value}
-                      onClick={() => handleCategory(option)}
-                    >
-                      {option.label}
-                    </Dropdown.Item>
-                  ))}
-                </DropdownButton>
-              </div>
-            </div>
-
-            <Modal.Body>
-              <Form.Control
-                className={styles.textArea}
-                as="textarea"
-                rows={5}
-                placeholder="What's on your mind?"
-                onChange={(event) => setContent(event.target.value)}
-              />
-            </Modal.Body>
-
-            <Modal.Footer>
-              <Button
-                className={styles.postBtn}
-                variant={`${!content ? "secondary" : "primary"}`}
-                disabled={!content.trim()}
-                onClick={onPost}
-              >
-                Post
-              </Button>
-            </Modal.Footer>
-          </Modal.Dialog>
-        </div>
+        <CreateEditPostModal
+          onPost={onPost}
+          content={{ data: content, setContent }}
+          privacy={{ data: privacy, setPrivacy }}
+          category={{ data: category, setCategory }}
+          setShowModal={setShowModal}
+          modalnfo={{
+            submitBtn: "Post",
+            title: "Create post",
+          }}
+        />
       )}
     </>
   );
