+++ post-app/src/components/CreatePost/CreatePostStyle.module.css
@@ -1,8 +1,3 @@
-.username a {
-  display: block;
-  width: 67px;
-}
-
 .createPost {
   display: flex;
   column-gap: 10px;
@@ -13,38 +8,17 @@
   border-radius: 10px;
 }
 
-.titleAndXButton {
-  display: grid;
-  grid-template-columns: 15fr 1fr;
-  column-gap: 20px;
-  justify-content: end;
-}
-
-.title {
-  justify-self: center;
-}
-
 .btnWhatsOnYourMind {
   min-width: 180px;
   width: 100%;
   height: 42px;
   border-radius: 20px;
   text-align: start;
-  
 }
 
-.postBtn {
-  width: 100%;
-}
-
-.userAndDropDown {
-  margin-left: 2px;
-  margin-bottom: 8px;
-}
-
-.userAndDropDowns {
-  display: flex;
-  justify-content: space-between;
+.username a {
+  display: block;
+  width: 67px;
 }
 
 .username {
@@ -63,28 +37,6 @@
   font-size: 18px;
 }
 
-.textArea {
-  margin-bottom: 20px;
-  margin-top: -2px;
-  background-color: rgb(217, 225, 233);
-  border: none;
-  font-size: 18px;
-}
-
-.createPostModal {
-  display: block;
-  margin: 40px 25px 0;
-  padding: 10px 15px;
-  position: initial;
-  background-color: rgb(73, 71, 71);
-  border-radius: 10px;
-  color: rgb(221, 225, 228);
-}
-
 p {
   color: rgb(228, 230, 235);
-}
-
-.dropdownButton {
-  margin-bottom: 20px;
 }
